requirements.txt
Pillow==4.3.0
pandas==0.21.0
matplotlib==2.1.0
MySQL-python==1.2.5
netCDF4==1.2.8
Pydap==3.2.2
redis==2.10.6
requests==2.18.4
simplejson==3.12.0
sorl-thumbnail==12.3
xlrd==1.1.0
Django==1.11.7
django-bootstrap3==9.1.0
django-debug-toolbar==1.9.1
django-extensions==1.9.7
django-filebrowser==3.9.1
django-grappelli==2.10.1
django-polymorphic==1.3
django-redis==4.8.0
django-registration==2.2
django-smart-selects==1.5.2

